# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation ",
        "description": "API Documentation",
        "contact": {
            "email": "srundavith@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get List movies Data",
                "description": "enter your movies here",
                "operationId": "movies",
                "responses": {
                    "default": {
                        "description": "return array model movies"
                    }
                }
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Create a new movie",
                "description": "Create a new movie with various attributes including file uploads for poster and cover images.",
                "operationId": "createMovie",
                "requestBody": {
                    "description": "Form data for creating a movie",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "overview",
                                    "run_time",
                                    "release_date",
                                    "total_likes",
                                    "total_ratings",
                                    "average_rating",
                                    "poster_image",
                                    "cover_image",
                                    "trailer_url",
                                    "last_upload_date",
                                    "subscription_only",
                                    "expired_subscription_only"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "overview": {
                                        "type": "string"
                                    },
                                    "run_time": {
                                        "type": "integer"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total_likes": {
                                        "type": "integer"
                                    },
                                    "total_ratings": {
                                        "type": "integer"
                                    },
                                    "average_rating": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "poster_image_file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "poster_image_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "cover_image_file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "cover_image_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "trailer_url": {
                                        "type": "string"
                                    },
                                    "last_upload_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "subscription_only": {
                                        "description": "Boolean represented as 1 (true) or 0 (false)",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "expired_subscription_only": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/api/tv_shows": {
            "get": {
                "tags": [
                    "Tv_Shows"
                ],
                "summary": "Get List Data",
                "description": "enter your  here",
                "operationId": "tv_shows",
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "post": {
                "tags": [
                    "Tv_Shows"
                ],
                "summary": "tv_shows",
                "description": "tv_shows",
                "operationId": "Tv_Shows",
                "requestBody": {
                    "description": "form tv_shows",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tv_show_name"
                                ],
                                "properties": {
                                    "tv_show_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}